python/pyproject.toml
# Copyright (c) 2023-2024, NVIDIA CORPORATION. All rights reserved.
# SPDX-License-Identifier: Apache-2.0

[project]
name = "legate-dataframe"
dynamic = ["version"]
authors = [
    {name = "NVIDIA Corporation"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
description = "Basic cudf operations on Legate"
license = {text = "Apache 2.0"}
requires-python = ">=3.10"
dependencies = [
    "cudf==24.08.*,>=0.0.0a0",
    "legate==24.11.01",
    "numpy >=1.23,<3.0.0a0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../dependencies.yaml and run `rapids-dependency-file-generator`.

[project.urls]
# TODO: Should have a homepage and docs (and add to meta.yaml!)
Repository = "https://github.com/rapidsai/legate-dataframe"
License = "https://github.com/rapidsai/legate-dataframe/blob/main/LICENSE"

[project.optional-dependencies]
test = [
    "cupynumeric==24.11.*",
    "dask-cuda==24.08.*",
    "dask-cudf==24.08.*",
    "pytest>=7.0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../dependencies.yaml and run `rapids-dependency-file-generator`.

[build-system]
requires = [
    "cmake>=3.26.4,!=3.30.0",
    "cython>=3.0.0",
    "legate==24.11.01",
    "libcudf==24.08.*,>=0.0.0a0",
    "ninja",
    "scikit-build>=0.18.0",
    "setuptools>=70.0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../dependencies.yaml and run `rapids-dependency-file-generator`.
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {file = "legate_dataframe/VERSION"}

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.py?$'
exclude = '''
/(
    thirdparty |
    \.eggs |
    \.git |
    \.hg |
    \.mypy_cache |
    \.tox |
    \.venv |
    _build |
    buck-out |
    build |
    dist |
    _skbuild
)/
'''

[tool.isort]
line_length = 88
profile = "black"
known_first_party = [
    "legate_dataframe",
]
known_cython = [
    "libc",
    "libcpp",
    "cython",
]
sections = [
    "FUTURE",
    "CYTHON",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
